Type Alias sshpuffin::libssh::ssh::AuthState

pub type AuthState = ssh_auth_state_e;

Aliased Type

enum AuthState {

    SSH_AUTH_STATE_NONE = 0,
    SSH_AUTH_STATE_PARTIAL = 1,
    SSH_AUTH_STATE_SUCCESS = 2,
    SSH_AUTH_STATE_FAILED = 3,
    SSH_AUTH_STATE_ERROR = 4,
    SSH_AUTH_STATE_INFO = 5,
    SSH_AUTH_STATE_PK_OK = 6,
    SSH_AUTH_STATE_KBDINT_SENT = 7,
    SSH_AUTH_STATE_GSSAPI_REQUEST_SENT = 8,
    SSH_AUTH_STATE_GSSAPI_TOKEN = 9,
    SSH_AUTH_STATE_GSSAPI_MIC_SENT = 10,
    SSH_AUTH_STATE_PUBKEY_OFFER_SENT = 11,
    SSH_AUTH_STATE_PUBKEY_AUTH_SENT = 12,
    SSH_AUTH_STATE_PASSWORD_AUTH_SENT = 13,
    SSH_AUTH_STATE_AUTH_NONE_SENT = 14,

}

Variants

SSH_AUTH_STATE_NONE = 0

No authentication asked

SSH_AUTH_STATE_PARTIAL = 1

Last authentication response was a partial success

SSH_AUTH_STATE_SUCCESS = 2

Last authentication response was a success

SSH_AUTH_STATE_FAILED = 3

Last authentication response was failed

SSH_AUTH_STATE_ERROR = 4

Last authentication was erroneous

SSH_AUTH_STATE_INFO = 5

Last state was a keyboard-interactive ask for info

SSH_AUTH_STATE_PK_OK = 6

Last state was a public key accepted for authentication

SSH_AUTH_STATE_KBDINT_SENT = 7

We asked for a keyboard-interactive authentication

SSH_AUTH_STATE_GSSAPI_REQUEST_SENT = 8

We have sent an userauth request with gssapi-with-mic

SSH_AUTH_STATE_GSSAPI_TOKEN = 9

We are exchanging tokens until authentication

SSH_AUTH_STATE_GSSAPI_MIC_SENT = 10

We have sent the MIC and expecting to be authenticated

SSH_AUTH_STATE_PUBKEY_OFFER_SENT = 11

We have offered a pubkey to check if it is supported

SSH_AUTH_STATE_PUBKEY_AUTH_SENT = 12

We have sent pubkey and signature expecting to be authenticated

SSH_AUTH_STATE_PASSWORD_AUTH_SENT = 13

We have sent a password expecting to be authenticated

SSH_AUTH_STATE_AUTH_NONE_SENT = 14

We have sent a request without auth information (method ‘none’)